test(jest): catch errors when calling `run()`…
…in order to prevent Jest to raise an `UnhandledPromiseRejectionWarning`

Ref #28
