+++ lib/utils.js
@@ -9,7 +9,7 @@ const path = require('path')
 const readPkgUp = require('read-pkg-up')
 
 // Used only for logging/debuggin
-var Table = require('cli-table')
+const Table = require('cli-table')
 
 // Match hunk headers.
 // Select previous line range with updated one to check
@@ -139,6 +139,54 @@ function notifyErrors({ message, kind }) {
     icon: path.join(__dirname, '..', 'assets', `checks-${kind}.png`),
     message,
   })
+}
+
+// Read staged files content and check if any pattern is matched.
+// If so, then list filenames for error and/or warnings.
+function parseContents({
+  display: { 'offending-content': printContent, verbose } = {},
+  pattern: { filter, nonBlocking, regex },
+  stagedContents,
+}) {
+  const errors = []
+  const warnings = []
+
+  for (const { content, fileName } of stagedContents) {
+    // Skip file if it does not match filter (and skip self)
+    if (filter && !filter.test(fileName)) {
+      continue
+    }
+
+    let matchFound = false
+    // Loop over each staged line in file
+    for (const [lineNumber, text] of content) {
+      if (!regex.test(text)) {
+        continue
+      }
+
+      if (verbose) {
+        matchFound = true
+        console.log(
+          `Match found in "${fileName}" at ${lineNumber} using regex "${regex}"`
+        )
+      }
+
+      const container = nonBlocking ? warnings : errors
+      // Use `filename:lineo` pattern for logging in order to
+      // enable quick-open on editors.
+      container.push(
+        `${fileName}:${lineNumber}${
+          printContent ? ' : ' + text.trimLeft() : ''
+        }`
+      )
+    }
+
+    if (!matchFound && verbose) {
+      console.log(`No match found in "${fileName}" using regex "${regex}"`)
+    }
+  }
+
+  return { errors, warnings }
 }
 
 // Group stage lines with their number.
@@ -305,6 +353,7 @@ module.exports = {
   largeOutputExec,
   loadPackageJSON,
   logAndExitWithTitle,
+  parseContents,
   parseStagedContent,
   printRulesSummary,
   printErrors,
