+++ package.json
@@ -166,6 +166,6 @@
   "dependencies": {
     "cli-table": "^0.3.1",
     "node-notifier": "^5.4.0",
-    "read-pkg-up": "^4.0.0"
+    "read-pkg-up": "^5.0.0"
   }
 }
